test: move tests and helpers to top-level test/ directory
Relocates all *.test.ts files and test helpers out of src/ so that
Sapphire's filesystem crawler (which scans src/ in dev mode via tsx
watch, and dist/ in production) never sees them as commands or
preconditions to load.

New layout:
  test/helpers/{interaction,run-command,sapphire}.ts
  test/utils/*.test.ts
  test/lib/*.test.ts
  test/preconditions/*.test.ts
  test/commands/compoverse/*.test.ts

Updates imports in every moved file, points vitest.config.ts at
test/**/*.test.ts, drops the now-unneeded test excludes from
tsconfig.json, and extends oxlint/oxfmt scripts to cover test/.
