+++ package.json
@@ -34,10 +34,10 @@
     "start": "node dist/index.js",
     "lint": "npm-run-all --aggregate-output --continue-on-error --parallel 'lint:!(fix)'",
     "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel 'lint:*:fix'",
-    "lint:oxlint": "oxlint src/",
-    "lint:oxlint:fix": "oxlint src/ --fix",
-    "lint:oxfmt": "oxfmt --check src/",
-    "lint:oxfmt:fix": "oxfmt src/",
+    "lint:oxlint": "oxlint src/ test/",
+    "lint:oxlint:fix": "oxlint src/ test/ --fix",
+    "lint:oxfmt": "oxfmt --check src/ test/",
+    "lint:oxfmt:fix": "oxfmt src/ test/",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage"
