+++ test/commands/compoverse/playlist.test.ts
@@ -1,22 +1,22 @@
 import { describe, it, expect, beforeAll, beforeEach, vi } from 'vitest';
 
-vi.mock('../../lib/party.js', () => ({
+vi.mock('../../../src/lib/party.js', () => ({
   partyService: {
     getSnapshot: vi.fn(),
   },
 }));
 
 import { MessageFlags } from 'discord.js';
-import { PlaylistCommand } from './playlist.js';
-import { CompoAdminOnly } from '../../preconditions/CompoAdminOnly.js';
-import { partyService } from '../../lib/party.js';
+import { PlaylistCommand } from '../../../src/commands/compoverse/playlist.js';
+import { CompoAdminOnly } from '../../../src/preconditions/CompoAdminOnly.js';
+import { partyService } from '../../../src/lib/party.js';
 import {
   makeAdminInteraction,
   makeNonAdminInteraction,
   makeMissingMemberInteraction,
-} from '../../__test_helpers__/interaction.js';
-import { registerForTest } from '../../__test_helpers__/sapphire.js';
-import { runCommand } from '../../__test_helpers__/run-command.js';
+} from '../../helpers/interaction.js';
+import { registerForTest } from '../../helpers/sapphire.js';
+import { runCommand } from '../../helpers/run-command.js';
 
 function makeSong(overrides: {
   id: string;
