+++ test/commands/compoverse/refetchparty.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeAll, beforeEach, vi } from 'vitest';
 
-vi.mock('../../lib/party.js', () => ({
+vi.mock('../../../src/lib/party.js', () => ({
   partyService: {
     getSnapshot: vi.fn(),
     send: vi.fn(),
@@ -8,24 +8,31 @@ vi.mock('../../lib/party.js', () => ({
 }));
 
 import { MessageFlags } from 'discord.js';
-import { StopPartyCommand } from './stopparty.js';
-import { CompoAdminOnly } from '../../preconditions/CompoAdminOnly.js';
-import { partyService } from '../../lib/party.js';
+import { RefetchPartyCommand } from '../../../src/commands/compoverse/refetchparty.js';
+import { CompoAdminOnly } from '../../../src/preconditions/CompoAdminOnly.js';
+import { partyService } from '../../../src/lib/party.js';
 import {
   makeAdminInteraction,
   makeNonAdminInteraction,
   makeMissingMemberInteraction,
-} from '../../__test_helpers__/interaction.js';
-import { registerForTest } from '../../__test_helpers__/sapphire.js';
-import { runCommand } from '../../__test_helpers__/run-command.js';
+} from '../../helpers/interaction.js';
+import { registerForTest } from '../../helpers/sapphire.js';
+import { runCommand } from '../../helpers/run-command.js';
 
-describe('StopPartyCommand', () => {
-  let command: StopPartyCommand;
+function matchesOnly(states: Array<string | Record<string, any>>) {
+  return (state: string | Record<string, any>) => {
+    const key = JSON.stringify(state);
+    return states.some((s) => JSON.stringify(s) === key);
+  };
+}
+
+describe('RefetchPartyCommand', () => {
+  let command: RefetchPartyCommand;
 
   beforeAll(async () => {
     command = await registerForTest({
       preconditions: [{ name: 'CompoAdminOnly', piece: CompoAdminOnly }],
-      command: { name: 'stopparty', piece: StopPartyCommand },
+      command: { name: 'refetchparty', piece: RefetchPartyCommand },
     });
   });
 
@@ -43,12 +50,10 @@ describe('StopPartyCommand', () => {
       (partyService.getSnapshot as any).mockReturnValue({ matches: () => false });
       const res = await runCommand(command, makeAdminInteraction());
       expect(res.ran).toBe(true);
-      expect(res.blockedBy).toBeNull();
     });
 
     it('blocks a non-admin member', async () => {
-      const interaction = makeNonAdminInteraction();
-      const res = await runCommand(command, interaction);
+      const res = await runCommand(command, makeNonAdminInteraction());
       expect(res.ran).toBe(false);
       expect(res.blockedBy).toBe('CompoAdminOnly');
       expect(partyService.send).not.toHaveBeenCalled();
@@ -63,9 +68,7 @@ describe('StopPartyCommand', () => {
 
   describe('chatInputRun', () => {
     it('replies ephemerally when idle', async () => {
-      (partyService.getSnapshot as any).mockReturnValue({
-        matches: (state: string) => state === 'idle',
-      });
+      (partyService.getSnapshot as any).mockReturnValue({ matches: matchesOnly(['idle']) });
       const interaction = makeAdminInteraction();
 
       const res = await runCommand(command, interaction);
@@ -73,24 +76,39 @@ describe('StopPartyCommand', () => {
       expect(res.ran).toBe(true);
       expect(partyService.send).not.toHaveBeenCalled();
       expect(interaction.reply).toHaveBeenCalledWith({
-        content: 'there is no listening party to stop!',
+        content: 'there is no listening party, currently!',
         flags: MessageFlags.Ephemeral,
       });
     });
 
-    it('sends STOP with immediate:true and replies when a party is running', async () => {
+    it('replies ephemerally when a fetch or refetch is already running', async () => {
+      (partyService.getSnapshot as any).mockReturnValue({ matches: () => false });
+      const interaction = makeAdminInteraction();
+
+      const res = await runCommand(command, interaction);
+
+      expect(res.ran).toBe(true);
+      expect(partyService.send).not.toHaveBeenCalled();
+      expect(interaction.reply).toHaveBeenCalledWith({
+        content: 'there is already a fetch or refetch running!',
+        flags: MessageFlags.Ephemeral,
+      });
+    });
+
+    it('sends REFETCH and replies when processing is idle', async () => {
       (partyService.getSnapshot as any).mockReturnValue({
-        matches: (state: string) => state !== 'idle',
+        matches: matchesOnly([{ partying: { processing: 'idle' } }]),
       });
       const interaction = makeAdminInteraction();
 
       const res = await runCommand(command, interaction);
 
       expect(res.ran).toBe(true);
-      expect(partyService.send).toHaveBeenCalledWith({ type: 'STOP', immediate: true });
-      expect(interaction.reply).toHaveBeenCalledWith({
-        content: 'Stopping the listening party...',
+      expect(partyService.send).toHaveBeenCalledWith({
+        type: 'REFETCH',
+        channel: interaction.channel,
       });
+      expect(interaction.reply).toHaveBeenCalledWith({ content: 'Refetching round...' });
     });
   });
 });
