+++ test/commands/compoverse/startparty.test.ts
@@ -1,29 +1,29 @@
 import { describe, it, expect, beforeAll, beforeEach, vi } from 'vitest';
 
-vi.mock('../../lib/party.js', () => ({
+vi.mock('../../../src/lib/party.js', () => ({
   partyService: {
     getSnapshot: vi.fn(),
     send: vi.fn(),
   },
 }));
 
-vi.mock('../../lib/logger.js', () => ({
+vi.mock('../../../src/lib/logger.js', () => ({
   log: vi.fn(),
   setDebugChannel: vi.fn(),
 }));
 
 import { MessageFlags } from 'discord.js';
-import { StartPartyCommand } from './startparty.js';
-import { CompoAdminOnly } from '../../preconditions/CompoAdminOnly.js';
-import { partyService } from '../../lib/party.js';
-import { log } from '../../lib/logger.js';
+import { StartPartyCommand } from '../../../src/commands/compoverse/startparty.js';
+import { CompoAdminOnly } from '../../../src/preconditions/CompoAdminOnly.js';
+import { partyService } from '../../../src/lib/party.js';
+import { log } from '../../../src/lib/logger.js';
 import {
   makeAdminInteraction,
   makeNonAdminInteraction,
   makeMissingMemberInteraction,
-} from '../../__test_helpers__/interaction.js';
-import { registerForTest } from '../../__test_helpers__/sapphire.js';
-import { runCommand } from '../../__test_helpers__/run-command.js';
+} from '../../helpers/interaction.js';
+import { registerForTest } from '../../helpers/sapphire.js';
+import { runCommand } from '../../helpers/run-command.js';
 
 describe('StartPartyCommand', () => {
   let command: StartPartyCommand;
