+++ test/utils/roles.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { memberHasOneOfTheseRoles, roleIds } from './roles.js';
+import { memberHasOneOfTheseRoles, roleIds } from '../../src/utils/roles.js';
 
 function makeMember(memberRoleIds: string[]) {
   return {
